+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 # Configuracion de la base de datos
 spring.datasource.url=jdbc:mysql://localhost:3306/flysky
 spring.datasource.username=root
-spring.datasource.password=pigush
+spring.datasource.password=Estud|4nd0
 #UVP70FI3lvE6MPbq
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
