+++ src/test/java/ar/com/codoacodo/flysky/integrationTest/IntegrationTestBoletos.java
@@ -11,13 +11,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
+import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.web.servlet.MockMvc;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-
+@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
 @AutoConfigureMockMvc
 @SpringBootTest
 public class IntegrationTestBoletos {
@@ -29,8 +30,7 @@ public class IntegrationTestBoletos {
         mockMvc.perform(get("/api/boletos/"))
                 .andDo(print())
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$[0].id").value(4))
-                .andExpect(jsonPath("$[0].nombre").value("Aerolíneas Argentinas"));
+                .andExpect(jsonPath("$[0].id").value(1));
     }
     @Test
     @DisplayName("Test Integracion Camino Feliz -> buscar un boleto por id")
@@ -39,8 +39,7 @@ public class IntegrationTestBoletos {
                 .andDo(print())
                 .andExpect(content().contentType("application/json"))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.id").value(2))
-                .andExpect(jsonPath("$.nombre").value("Avianca"));
+                .andExpect(jsonPath("$.id").value(2));
     }
     @Test
     @DisplayName("Test Integracion Camino Feliz -> Agregar un nuevo boleto")
