+++ src/test/java/ar/com/codoacodo/flysky/integrationTestException/ExceptionsIntegrationTestAerolineas.java
@@ -1,4 +1,4 @@
-package ar.com.codoacodo.flysky.integrationTest;
+package ar.com.codoacodo.flysky.integrationTestException;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -6,39 +6,41 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
+import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.web.servlet.MockMvc;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
 @AutoConfigureMockMvc
 @SpringBootTest
-public class ExceptionsIntegrationTestPagos {
+public class ExceptionsIntegrationTestAerolineas {
     @Autowired
     MockMvc mockMvc;
-    
-    @Test //Aeronaves
-    @DisplayName("Test Integracion Camino Feliz -> Integracion pagosException -> buscarPorId")
-    void pagosbuscarPorIdExceptionOkTest() throws Exception {
-        mockMvc.perform(get("/api/pagos/{id}", 188)
+    @Test //Aerolinas
+    @DisplayName("Test Integracion Camino Feliz -> Integracion AerolineasException -> buscarPorId")
+    void aerolineasbuscarPorIdExceptionOkTest() throws Exception {
+        mockMvc.perform(get("/api/aerolineas/{id}", 188)
                         .contentType(MediaType.APPLICATION_JSON))
                 .andDo(print())
                 .andExpect(status().isNotFound())
                 .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-                .andExpect(jsonPath("$.message").value("¡El pago solicitado no existe!"))
+                .andExpect(jsonPath("$.message").value("¡La aerolínea especificada no existe!"))
                 .andExpect(jsonPath("$.status").value(404));
     }
-    @Test //Aeronaves
-    @DisplayName("Test Integracion Camino Feliz -> Integracion pagosException -> eliminar")
-    void pagosEliminarExceptionOkTest() throws Exception {
-        mockMvc.perform(delete("/api/pagos/{id}/eliminar", 188)
+    @Test //Aerolinas
+    @DisplayName("Test Integracion Camino Feliz -> Integracion AerolineasException -> eliminar")
+    void aerolineasEliminarExceptionOkTest() throws Exception {
+        mockMvc.perform(delete("/api/aerolineas/{id}/eliminar", 188)
                         .contentType(MediaType.APPLICATION_JSON))
                 .andDo(print())
                 .andExpect(status().isNotFound())
                 .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-                .andExpect(jsonPath("$.message").value("¡El pago solicitado no existe!"))
+                .andExpect(jsonPath("$.message").value("¡La aerolínea especificada no existe!"))
                 .andExpect(jsonPath("$.status").value(404));
     }
+
 }
